+++ src/PageProtector.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\DataModel\Page;
+
+class PageProtector {
+
+	/**
+	 * @since 0.3
+	 *
+	 * @param Page $page
+	 * @todo Add options for different protection levels
+	 */
+	public function protect( Page $page ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}