+++ src/PageRestorer.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\DataModel\Page;
+
+class PageRestorer {
+
+	/**
+	 * @since 0.3
+	 *
+	 * @param Page $page
+	 */
+	public function restore( Page $page ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}