+++ src/RevisionRestorer.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\DataModel\Revision;
+
+class RevisionRestorer {
+
+	/**
+	 * @since 0.3
+	 *
+	 * @param Revision $revision
+	 */
+	public function restore( Revision $revision ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}