+++ src/UserBlocker.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\DataModel\User;
+
+class UserBlocker {
+
+	/**
+	 * @since 0.3
+	 *
+	 * @param User $user
+	 */
+	public function patrol( User $user ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}