+++ src/UserRightsChanger.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Mediawiki\Api\Service;
+
+use Mediawiki\DataModel\User;
+
+class UserRightsChanger {
+
+	/**
+	 * @since 0.3
+	 *
+	 * @param User $user
+	 * @param $rights
+	 */
+	public function change( User $user, $rights ) {
+		//TODO implement me
+		throw new \BadMethodCallException( 'Not yet implemented' );
+	}
+
+}